+++ aws/build.sh
@@ -5,7 +5,7 @@ source "/opt/sdaccel-builder/settings.sh"
 curl -XPOST -H "Content-Type: application/json"  -d '{"status": "STARTED"}' "$CALLBACK_URL" &> /dev/null
 
 set +e
-aws s3 cp --quiet "$INPUT_URL" - | tar zxf -
+aws s3 cp --quiet "$INPUT_URL" - | tar zxf - --transform='s,\\,/,g' --show-transformed-names
 
 exit="$?"